+++ src/Exception/MutexReleaseException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Mutex\Exception;
+
+use RuntimeException;
+
+final class MutexReleaseException extends RuntimeException
+{
+}